unix1
Write a script called recycle that mimics the rm command. The recycle script should be able to accept the name of a file as a command line argument as rm does, but instead of deleting the file, your script should move it to a recyclebin directory called recyclebin in your home directory.

1.	The script name is recycle and will be stored in $HOME/project. 
# cd in $HOME/project
# vi recycle // new script file name recycle/cy.
2.	The recycle bin will be $HOME/recyclebin. Ensure that your script creates this directory

#mkdir $HOME/recyclebin

3.	The file to be deleted will be a command line argument and the script should be executed as follows: 	bash recycle fileName 
#virable $fileName

4.	The script must test for the following error conditions and display the same kind of error messages as the rm command.

-	No filename provided - Display an error message if no filename is provided as an argument, and set an error exit status.
# if no filename 
	Return “”
	Exit

-	File does not exist - Display an error message if file supplied does not exist, and terminate the script.
 # if no filename 
	Return “”
	Exit



-	Directory name provided - Display an error message if a directory name is provided instead of a filename, and terminate the script.

-	Test that the file being deleted is not the recycle script. If it is, display the error message “Attempting to delete recycle – operation aborted” and terminate the script. Be sure to create a hard link for your script before testing that this works.

5.	The filenames in the recyclebin, will be in the following format:
fileName_inode
For example, if a file named f1 with inode 1234 is recycled, the file in the recyclebin will be named f1_1234. This gets around the potential problem of deleting two files with the same name. The recyclebin will only contain files, not directories. 

6.	Create a hidden file called .restore.info in $HOME.  Each line of this file will contain the name of the file in the recyclebin, followed by a colon, followed by the original absolute path of the file.  For example, if a file called f1, with an inode of 1234 was recycled from the /home/trainee1 directory, .restore.info will contain:
f1_1234:/home/trainee1/f1
If  another file named f1, with an inode of 5432, was recycled from the /home/trainee1/testing directory, then .restore.info will contain:
f1_1234:/home/trainee1/f1
f1_5432:/home/trainee1/testing/f1